+++ mcnc_2024/src/components/auth/sign-up-step/SignUpStep2.vue
